requirements.txt
beautifulsoup4==4.13.4
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
cssselect==1.3.0
idna==3.10
lxml==6.0.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.19.2
readability-lxml==0.8.4.1
requests==2.32.4
rich==14.1.0
soupsieve==2.7
termcolor==3.1.0
typing_extensions==4.14.1
urllib3==2.5.0
